git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@4598 406a0c4d-033a-0410-8de8-e80135713968

+++ src/sphinxexpr.cpp
@@ -1380,8 +1380,8 @@ DECLARE_UNARY_TRAITS ( Expr_Ln_c )
        virtual float Eval ( const CSphMatch & tMatch ) const
        {
                float fFirst = m_pFirst->Eval ( tMatch );
-               // ideally this would be SQLNULL instead of plain 0
-               return fFirst>0 ? log ( fFirst ) : 0;
+               // ideally this would be SQLNULL instead of plain 0.0f
+               return fFirst>0.0f ? log ( fFirst ) : 0.0f;
        }
 DECLARE_END()
 
@@ -1389,8 +1389,8 @@ DECLARE_UNARY_TRAITS ( Expr_Log2_c )
        virtual float Eval ( const CSphMatch & tMatch ) const
        {
                float fFirst = m_pFirst->Eval ( tMatch );
-               // ideally this would be SQLNULL instead of plain 0
-               return fFirst>0 ? log ( fFirst )*M_LOG2E : 0;
+               // ideally this would be SQLNULL instead of plain 0.0f
+               return fFirst>0.0f ? log ( fFirst )*M_LOG2E : 0.0f;
        }
 DECLARE_END()
 
@@ -1398,8 +1398,8 @@ DECLARE_UNARY_TRAITS ( Expr_Log10_c )
        virtual float Eval ( const CSphMatch & tMatch ) const
        {
                float fFirst = m_pFirst->Eval ( tMatch );
-               // ideally this would be SQLNULL instead of plain 0
-               return fFirst>0 ? log ( fFirst )*M_LOG10E : 0;
+               // ideally this would be SQLNULL instead of plain 0.0f
+               return fFirst>0.0f ? log ( fFirst )*M_LOG10E : 0.0f;
        }
 DECLARE_END()
 
@@ -1407,9 +1407,9 @@ DECLARE_UNARY_TRAITS ( Expr_Sqrt_c )
        virtual float Eval ( const CSphMatch & tMatch ) const
        {
                float fFirst = m_pFirst->Eval ( tMatch );
-               // ideally this would be SQLNULL instead of plain 0 in case of negative argument
-               // MEGA optimization: do not call sqrt for 0
-               return fFirst>0 ? sqrt ( fFirst ) : 0;
+               // ideally this would be SQLNULL instead of plain 0.0f in case of negative argument
+               // MEGA optimization: do not call sqrt for 0.0f
+               return fFirst>0.0f ? sqrt ( fFirst ) : 0.0f;
        }
 DECLARE_END()
 
@@ -1455,8 +1455,8 @@ DECLARE_BINARY_TRAITS ( Expr_Div_c )
        virtual float Eval ( const CSphMatch & tMatch ) const
        {
                float fSecond = m_pSecond->Eval ( tMatch );
-               // ideally this would be SQLNULL instead of plain 0
-               return fSecond ? m_pFirst->Eval ( tMatch )/fSecond : 0;
+               // ideally this would be SQLNULL instead of plain 0.0f
+               return fSecond ? m_pFirst->Eval ( tMatch )/fSecond : 0.0f;
        }
 DECLARE_END()
 
@@ -1464,7 +1464,7 @@ DECLARE_BINARY_TRAITS ( Expr_Idiv_c )
 	virtual float Eval ( const CSphMatch & tMatch ) const
 	{
 		int iSecond = int(SECOND);
-		// ideally this would be SQLNULL instead of plain 0
+		// ideally this would be SQLNULL instead of plain 0.0f
 		return iSecond ? float(int(FIRST)/iSecond) : 0.0f;
 	}
 
@@ -1546,7 +1546,7 @@ DECLARE_TERNARY ( Expr_Mul3_c,	FIRST*SECOND*THIRD,					INTFIRST*INTSECOND*INTTHI
 
 #define DECLARE_TIMESTAMP(_classname,_expr) \
 	DECLARE_UNARY_TRAITS ( _classname ) \
-		virtual float Eval ( const CSphMatch & tMatch ) const { return IntEval(tMatch); } \
+		virtual float Eval ( const CSphMatch & tMatch ) const { return (float)IntEval(tMatch); } \
 		virtual int64_t Int64Eval ( const CSphMatch & tMatch ) const { return IntEval(tMatch); } \
 		virtual int IntEval ( const CSphMatch & tMatch ) const \
 		{ \
@@ -2614,7 +2614,7 @@ void ExprParser_t::ConstantFoldPass ( int iNode )
 					case '+':	pRoot->m_fConst = fLeft + fRight; break;
 					case '-':	pRoot->m_fConst = fLeft - fRight; break;
 					case '*':	pRoot->m_fConst = fLeft * fRight; break;
-					case '/':	pRoot->m_fConst = fRight ? fLeft / fRight : 0; break;
+					case '/':	pRoot->m_fConst = fRight ? fLeft / fRight : 0.0f; break;
 					default:	assert ( 0 && "internal error: unhandled arithmetic token during const-float optimization" );
 				}
 				pRoot->m_iToken = TOK_CONST_FLOAT;
@@ -2692,11 +2692,11 @@ void ExprParser_t::ConstantFoldPass ( int iNode )
 			case FUNC_FLOOR:	pRoot->m_iToken = TOK_CONST_INT; pRoot->m_iLeft = -1; pRoot->m_iConst = (int)floor ( fArg ); break;
 			case FUNC_SIN:		pRoot->m_iToken = TOK_CONST_FLOAT; pRoot->m_iLeft = -1; pRoot->m_fConst = float ( sin ( fArg) ); break;
 			case FUNC_COS:		pRoot->m_iToken = TOK_CONST_FLOAT; pRoot->m_iLeft = -1; pRoot->m_fConst = float ( cos ( fArg ) ); break;
-			case FUNC_LN:		pRoot->m_iToken = TOK_CONST_FLOAT; pRoot->m_iLeft = -1; pRoot->m_fConst = fArg>0 ? log(fArg) : 0; break;
-			case FUNC_LOG2:		pRoot->m_iToken = TOK_CONST_FLOAT; pRoot->m_iLeft = -1; pRoot->m_fConst = fArg>0 ? log(fArg)*M_LOG2E : 0; break;
-			case FUNC_LOG10:	pRoot->m_iToken = TOK_CONST_FLOAT; pRoot->m_iLeft = -1; pRoot->m_fConst = fArg>0 ? log(fArg)*M_LOG10E : 0; break;
+			case FUNC_LN:		pRoot->m_iToken = TOK_CONST_FLOAT; pRoot->m_iLeft = -1; pRoot->m_fConst = fArg>0.0f ? log(fArg) : 0.0f; break;
+			case FUNC_LOG2:		pRoot->m_iToken = TOK_CONST_FLOAT; pRoot->m_iLeft = -1; pRoot->m_fConst = fArg>0.0f ? log(fArg)*M_LOG2E : 0.0f; break;
+			case FUNC_LOG10:	pRoot->m_iToken = TOK_CONST_FLOAT; pRoot->m_iLeft = -1; pRoot->m_fConst = fArg>0.0f ? log(fArg)*M_LOG10E : 0.0f; break;
 			case FUNC_EXP:		pRoot->m_iToken = TOK_CONST_FLOAT; pRoot->m_iLeft = -1; pRoot->m_fConst = float ( exp ( fArg ) ); break;
-			case FUNC_SQRT:		pRoot->m_iToken = TOK_CONST_FLOAT; pRoot->m_iLeft = -1; pRoot->m_fConst = fArg>0 ? sqrt(fArg) : 0; break;
+			case FUNC_SQRT:		pRoot->m_iToken = TOK_CONST_FLOAT; pRoot->m_iLeft = -1; pRoot->m_fConst = fArg>0.0f ? sqrt(fArg) : 0.0f; break;
 			default:			break;
 		}
 		return;
